+++ src/Streamly.hs
@@ -224,7 +224,7 @@ where
 
 import Data.Semigroup (Semigroup(..))
 import Streamly.Internal.Data.SVar (MonadAsync, Rate(..))
-import Streamly.Streams.Ahead
+import Streamly.Internal.Data.Stream.Ahead
 import Streamly.Streams.Async
 import Streamly.Streams.Combinators
 import Streamly.Streams.Parallel
